Clown
This program solves the LSAT clown puzzle where 4 of the 7 clowns
participating in the circus are chosen to perform, yet, certain
conditions must be met before placing them in a group.

Link to the problem is found here:
https://www.cambridgelsat.com/assets/pdfs/logic-games-practice/Circus-Ga
me-(15).pdf

Requires Racket Version 6.3 to run program.

By inputting the list of clowns to choose from in the interactions
window, the program will give you back a list of remaining clowns you
may choose.

Inputting the clowns in the following format: (make-chst (list)(list M
N O P R S V)), with the variables replaced with the defined variables
written on the top of the program, will give desired output.